requirements.txt
asgiref==3.8.1
certifi==2025.4.26
charset-normalizer==3.4.2
cloudinary==1.44.1
confusable-homoglyphs==3.3.1
dj-database-url==3.0.1
Django==5.2.3
django-cors-headers==4.7.0
django-registration==5.2.1
django-rest-framework==0.1.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-yasg==1.21.10
gunicorn==23.0.0
idna==3.10
inflection==0.5.1
Markdown==3.8.2
packaging==25.0
psycopg2==2.9.10
psycopg2-binary==2.9.10
PyJWT==2.9.0
PyPDF2==3.0.1
python-decouple==3.8
pytz==2025.2
PyYAML==6.0.2
requests==2.32.4
six==1.17.0
sqlparse==0.5.3
uritemplate==4.2.0
urllib3==2.5.0
